fixed import after split of feature extraction
