+++ asoid/utils/extract_features_2D.py
@@ -186,7 +186,7 @@ def feature_extraction(train_datalist, num_train, framerate):
     # for i in stqdm(range(num_train), desc="Extracting spatiotemporal features from pose"):
     for i in range(num_train):
         data_list = List()
-        with st.spinner('Extracting features from pose...'):
+        with st.spinner('Extracting features from 2D pose...'):
             data_list.append(train_datalist[i])
             binned_features = bsoid_extract_numba(data_list, framerate)
             f_integrated.append(binned_features[0])  # getting only the non-shifted
